+++ spec/integration/braintree/dispute_search_spec.rb
@@ -107,6 +107,29 @@ describe Braintree::Dispute, "search" do
       expect(dispute.reason).to eq(Braintree::Dispute::Reason::Fraud)
     end
 
+    context "pre-dispute program" do
+      it "correctly returns disputes by pre-dispute program" do
+        collection = Braintree::Dispute.search do |search|
+          search.pre_dispute_program.in [
+            Braintree::Dispute::PreDisputeProgram::VisaRdr
+          ]
+        end
+
+        expect(collection.disputes.count).to eq(1)
+        dispute = collection.disputes.first
+        expect(dispute.pre_dispute_program).to eq(Braintree::Dispute::PreDisputeProgram::VisaRdr)
+      end
+
+      it "correctly returns disputes with no pre-dispute program" do
+        collection = Braintree::Dispute.search do |search|
+          search.pre_dispute_program.is Braintree::Dispute::PreDisputeProgram::None
+        end
+
+        expect(collection.disputes.count).to be > 1
+        expect(collection.disputes.map(&:pre_dispute_program).uniq).to eq([Braintree::Dispute::PreDisputeProgram::None])
+      end
+    end
+
     it "correctly returns disputes by effective_date range" do
       effective_date = transaction.disputes.first.status_history.first.effective_date